Add popen3 so that dirname doesn't pollute stderr
Previously, when Mobot_dongleGetTTY() failed because no dongle is
plugged in, dirname polluted stderr with its own failure messages. Since
these errors are both expected in this case and reflect
Mobot_dongleGetTTY's implementation details, it would be better if the
child process' stderr is suppressed. Since popen() couldn't do this, I
wrote popen3(), which automatically captures stdin, stdout, and stderr
for the calling process to do stuff with.

In libbarobo's case, it doesn't have anything to do, so it ignores them.