Pass -DBAROBO_BUILD_CHMOBOT to build chmobot now
Previously, chmobot built on linux if CHHOME was defined. Now, it is
built if you set BAROBO_BUILD_CHMOBOT to ON or TRUE or whatever on the
command line or through ccmake. The only difference between the two
modes is that the barobo target is static for chmobot, and chmobot_src
is added as a subdirectory.